squash commits for elicitation support
