+++ packages/insomnia/src/common/mcp-utils.ts
@@ -138,11 +138,11 @@ export const getMcpMethodFromMessage = (message: JSONRPCMessage): McpMessageEven
     }
   } else if (ServerRequestSchema.safeParse(message).success) {
     const requestMethod = ServerRequestSchema.parse(message).method;
-    if (requestMethod === METHOD_LIST_ROOTS) {
-      method = METHOD_LIST_ROOTS;
+    if (requestMethod === METHOD_LIST_ROOTS || requestMethod === METHOD_ELICITATION_CREATE_MESSAGE) {
+      method = requestMethod;
     } else {
       // Do not support any server requests to client including ping, elicitation and sampling
-      method = `${unsupportedMethodPrefix}${ServerRequestSchema.parse(message).method}`;
+      method = `${unsupportedMethodPrefix}${requestMethod}`;
     }
   }
   return method;
