+++ packages/insomnia/src/ui/components/mcp/mcp-notification-tab.tsx
@@ -2,7 +2,7 @@ import { useMemo, useState } from 'react';
 import { Button, Input, SearchField } from 'react-aria-components';
 import { Panel, PanelGroup, PanelResizeHandle } from 'react-resizable-panels';
 
-import type { McpNotificationEvent } from '~/main/network/mcp';
+import type { McpNotificationEvent } from '~/main/mcp/types';
 import { Icon } from '~/ui/components/icon';
 import { McpEventView } from '~/ui/components/mcp/event-view';
 import { EventLogView } from '~/ui/components/websockets/event-log-view';
