+++ packages/insomnia/src/ui/components/mcp/mcp-pane.tsx
@@ -28,7 +28,7 @@ import {
   METHOD_LIST_TOOLS,
 } from '~/common/mcp-utils';
 import { fuzzyMatchAll } from '~/common/misc';
-import type { McpEvent, McpMessageEvent } from '~/main/network/mcp';
+import type { McpEvent, McpMessageEvent } from '~/main/mcp/types';
 import type { McpRequest, McpServerPrimitiveTypes } from '~/models/mcp-request';
 import { useRootLoaderData } from '~/root';
 import { useWorkspaceLoaderData } from '~/routes/organization.$organizationId.project.$projectId.workspace.$workspaceId';
