+++ packages/insomnia/src/ui/components/mcp/mcp-request-pane.tsx
@@ -11,7 +11,6 @@ import type { McpReadyState } from '~/main/network/mcp';
 import { useWorkspaceLoaderData } from '~/routes/organization.$organizationId.project.$projectId.workspace.$workspaceId';
 import { Link } from '~/ui/components/base/link';
 import { EnvironmentKVEditor } from '~/ui/components/editors/environment-key-value-editor/key-value-editor';
-import { Icon } from '~/ui/components/icon';
 import { InsomniaRjsfForm, type InsomniaRjsfFormHandle } from '~/ui/components/rjsf';
 
 import { type AuthTypes } from '../../../common/constants';
@@ -67,7 +66,6 @@ export const McpRequestPane: FC<Props> = ({
 }) => {
   const primitiveId = `${selectedPrimitiveItem?.type}_${selectedPrimitiveItem?.name}`;
   const { activeRequest, activeRequestMeta, requestPayload } = useRequestLoaderData()! as McpRequestLoaderData;
-  const [isCalling, setIsCalling] = useState(false);
   const latestRequestPayloadRef = useLatest(requestPayload);
 
   const { activeProject } = useWorkspaceLoaderData()!;
@@ -135,36 +133,34 @@ export const McpRequestPane: FC<Props> = ({
   );
 
   const handleSend = async () => {
-    rjsfFormRef.current?.validate();
-    try {
-      setIsCalling(true);
-      if (selectedPrimitiveItem?.type === 'tools') {
-        await window.main.mcp.primitive.callTool({
-          name: selectedPrimitiveItem?.name || '',
-          arguments: mcpParams[primitiveId],
-          requestId: requestId,
-        });
-      } else if (selectedPrimitiveItem?.type === 'resources') {
-        await window.main.mcp.primitive.readResource({
-          requestId,
-          uri: selectedPrimitiveItem?.uri || '',
-        });
-      } else if (selectedPrimitiveItem?.type === 'resourceTemplates') {
-        await window.main.mcp.primitive.readResource({
-          requestId,
-          uri: fillUriTemplate(selectedPrimitiveItem.uriTemplate, mcpParams[primitiveId] || {}),
-        });
-      } else if (selectedPrimitiveItem?.type === 'prompts') {
-        await window.main.mcp.primitive.getPrompt({
-          requestId,
-          name: selectedPrimitiveItem?.name || '',
-          arguments: mcpParams[primitiveId],
-        });
+    if (rjsfFormRef.current?.validate()) {
+      try {
+        if (selectedPrimitiveItem?.type === 'tools') {
+          await window.main.mcp.primitive.callTool({
+            name: selectedPrimitiveItem?.name || '',
+            arguments: mcpParams[primitiveId],
+            requestId: requestId,
+          });
+        } else if (selectedPrimitiveItem?.type === 'resources') {
+          await window.main.mcp.primitive.readResource({
+            requestId,
+            uri: selectedPrimitiveItem?.uri || '',
+          });
+        } else if (selectedPrimitiveItem?.type === 'resourceTemplates') {
+          await window.main.mcp.primitive.readResource({
+            requestId,
+            uri: fillUriTemplate(selectedPrimitiveItem.uriTemplate, mcpParams[primitiveId] || {}),
+          });
+        } else if (selectedPrimitiveItem?.type === 'prompts') {
+          await window.main.mcp.primitive.getPrompt({
+            requestId,
+            name: selectedPrimitiveItem?.name || '',
+            arguments: mcpParams[primitiveId],
+          });
+        }
+      } catch (err) {
+        console.warn('MCP primitive call error', err);
       }
-    } catch (err) {
-      console.warn('MCP primitive call error', err);
-    } finally {
-      setIsCalling(false);
     }
   };
 
@@ -308,7 +304,6 @@ export const McpRequestPane: FC<Props> = ({
                           onClick={handleSend}
                           className="rounded bg-[--color-surprise] px-[--padding-md] text-center text-[--color-font-surprise]"
                         >
-                          {isCalling && <Icon className="mr-1 animate-spin" icon="spinner" />}
                           {sendButtonText}
                         </Button>
                       </div>
